fix request wrappers
